refactor(handlers): modify function to get full user from database
